[Core] Split locker listener to locking and simple product locking listeners
