+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/listeners.xml
@@ -71,11 +71,15 @@
             <tag name="kernel.event_listener" event="sylius.product_review.pre_create" method="ensureReviewHasAuthor" />
         </service>
 
-        <service id="sylius.listener.locker" class="Sylius\Bundle\CoreBundle\EventListener\LockerListener">
+        <service id="sylius.listener.locking" class="Sylius\Bundle\CoreBundle\EventListener\LockingListener">
+            <argument type="service" id="doctrine.orm.entity_manager" />
+            <tag name="kernel.event_listener" event="sylius.product_variant.initialize_update" method="lock" />
+        </service>
+
+        <service id="sylius.listener.simple_product_locking" class="Sylius\Bundle\CoreBundle\EventListener\SimpleProductLockingListener">
             <argument type="service" id="doctrine.orm.entity_manager" />
             <argument type="service" id="sylius.product_variant_resolver.default" />
             <tag name="kernel.event_listener" event="sylius.product.initialize_update" method="lock" />
-            <tag name="kernel.event_listener" event="sylius.product_variant.initialize_update" method="lock" />
         </service>
     </services>
 </container>
